refactor: ApiMain 클래스 이동 (#64)
