+++ api/src/main/kotlin/com/few/api/ApiMain.kt
@@ -1,4 +1,4 @@
-package com.few
+package com.few.api
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
